requirements.txt
flask
gunicorn
torch
facenet-pytorch
scikit-learn
pillow==12.0.0
numpy
joblib
python-dotenv
tqdm
matplotlib